+++ Web Ui/src/sections/Login/cookieService.tsx
@@ -2,7 +2,7 @@ import Cookies from 'js-cookie';
 
 export const setSessionCookie = (userData: any) => {
   try {
-    Cookies.set('userSession', JSON.stringify(userData), { expires: 100000 });
+    Cookies.set('userSession', JSON.stringify(userData), { expires: 1 });
     console.log('Session cookie set successfully.');
     console.log(userData);
   } catch (error) {
